Move start_flow to utils and task to sidekick
